+++ python_ta/transforms/ExprWrapper.py
@@ -1,8 +1,9 @@
-from typing import Dict, List, Optional, Set, Tuple, Union
+from typing import Dict, List, Optional, Union
 
 import astroid
 import z3
 from astroid import nodes
+from pylint.checkers.utils import safe_infer
 
 
 class Z3ParseException(Exception):
@@ -20,17 +21,22 @@ class ExprWrapper:
 
     Instance attributes:
         - node: astroid node obtained given by the value attribute of astroid expression.
-        - types: dictionary mapping variable names in astroid expression to their type name.
+        - types: dictionary mapping variable names in astroid expression to their type name or z3 variable.
     """
 
     node: astroid.NodeNG
-    types: Dict[str, str]
+    types: Dict[str, Union[str, z3.ExprRef]]
 
-    def __init__(self, expr: nodes.Expr, types=None):
-        self.node = expr.value
+    def __init__(self, node: astroid.NodeNG, types=None):
         if types is None:
             types = {}
         self.types = types
+
+        # extract expression out of expression statement
+        if isinstance(node, (astroid.Expr, astroid.Assign)):
+            self.node = node.value
+        else:
+            self.node = node
 
     def reduce(self, node: astroid.NodeNG = None) -> z3.ExprRef:
         """
@@ -52,8 +58,12 @@ class ExprWrapper:
             node = node.value
         elif isinstance(node, nodes.Name):
             node = self.apply_name(node.name)
+        elif isinstance(node, nodes.AssignName):
+            node = self.apply_name(node.name)
         elif isinstance(node, (nodes.List, nodes.Tuple, nodes.Set)):
             node = self.parse_container_op(node)
+        elif isinstance(node, nodes.Subscript):
+            node = self.parse_subscript_op(node)
         else:
             raise Z3ParseException(f"Unhandled node type {type(node)}.")
 
@@ -64,14 +74,17 @@ class ExprWrapper:
         Set up the appropriate variable representation in Z3 based on name and type.
         If an error is encountered or a case is unconsidered, return None.
         """
-        typ = self.types[name]
+        typ = self.types.get(name)
         type_to_z3 = {
             "int": z3.Int,
             "float": z3.Real,
             "bool": z3.Bool,
+            "str": z3.String,
         }
-        if typ in type_to_z3:
+        if typ in type_to_z3:  # convert string value to z3 variable
             x = type_to_z3[typ](name)
+        elif isinstance(typ, z3.ExprRef):  # the value is already a z3 variable
+            x = typ
         else:
             raise Z3ParseException(f"Unhandled type {typ}.")
 
@@ -116,6 +129,8 @@ class ExprWrapper:
                 return left**right
             elif op == "==":
                 return left == right
+            elif op == "!=":
+                return left != right
             elif op == "<=":
                 return left <= right
             elif op == ">=":
@@ -124,10 +139,10 @@ class ExprWrapper:
                 return left < right
             elif op == ">":
                 return left > right
-            elif op == "in" and isinstance(right, list):
-                return z3.Or(*[left == element for element in right])
-            elif op == "not in" and isinstance(right, list):
-                return z3.And(*[left != element for element in right])
+            elif op == "in":
+                return self.apply_in_op(left, right)
+            elif op == "not in":
+                return self.apply_in_op(left, right, negate=True)
             else:
                 raise Z3ParseException(
                     f"Unhandled binary operation {op} with operator types {left} and {right}."
@@ -153,7 +168,6 @@ class ExprWrapper:
         """Convert an astroid UnaryOp node to a z3 expression."""
         left, op = node.operand, node.op
         left = self.reduce(left)
-
         return self.apply_unary_op(left, op)
 
     def parse_bin_op(self, node: astroid.BinOp) -> z3.ExprRef:
@@ -172,7 +186,119 @@ class ExprWrapper:
         return self.apply_bool_op(op, values)
 
     def parse_container_op(
-        self, node: Union[nodes.List, nodes.Set, nodes.Tuple]
+        self, node: Union[nodes.List, astroid.Set, astroid.Tuple]
     ) -> List[z3.ExprRef]:
         """Convert an astroid List, Set, Tuple node to a list of z3 expressions."""
         return [self.reduce(element) for element in node.elts]
+
+    def apply_in_op(
+        self,
+        left: Union[z3.ExprRef, str],
+        right: Union[z3.ExprRef, List[z3.ExprRef], str],
+        negate: bool = False,
+    ) -> z3.ExprRef:
+        """
+        Apply `in` or `not in` operator on a list or string and return the
+        resulting z3 expression. Raise Z3ParseException if the operands
+        do not support `in` operator
+        """
+        if isinstance(right, list):  # container type (list/set/tuple)
+            return (
+                z3.And(*[left != element for element in right])
+                if negate
+                else z3.Or(*[left == element for element in right])
+            )
+        elif isinstance(left, (str, z3.SeqRef)) and isinstance(
+            right, (str, z3.SeqRef)
+        ):  # string literal or variable
+            return z3.Not(z3.Contains(right, left)) if negate else z3.Contains(right, left)
+        else:
+            op = "not in" if negate else "in"
+            raise Z3ParseException(
+                f"Unhandled binary operation {op} with operator types {left} and {right}."
+            )
+
+    def _parse_number_literal(self, node: astroid.NodeNG) -> Optional[Union[int, float]]:
+        """
+        If the subtree from `node` represent a number literal, return the value
+        Otherwise, return None
+        """
+        # positive number
+        if isinstance(node, nodes.Const) and isinstance(node.value, (int, float)):
+            return node.value
+        # negative number
+        elif (
+            isinstance(node, nodes.UnaryOp)
+            and node.op == "-"
+            and isinstance(node.operand, nodes.Const)
+            and isinstance(node.operand.value, (int, float))
+        ):
+            return -node.operand.value
+        else:
+            return None
+
+    def parse_subscript_op(self, node: astroid.Subscript) -> z3.ExprRef:
+        """
+        Convert an astroid Subscript node to z3 expression.
+        This method only supports string values and integer literal (both positive and negative) indexes
+        """
+        value = self.reduce(node.value)
+        slice = node.slice
+
+        # check for invalid node type
+        if not z3.is_seq(value):
+            raise Z3ParseException(f"Unhandled subscript operand type {value}")
+
+        # handle indexing
+        index = self._parse_number_literal(slice)
+        if isinstance(index, int):
+            return z3.SubString(value, index, 1)
+
+        # handle slicing
+        if isinstance(slice, nodes.Slice):
+            lower = 0 if slice.lower is None else self._parse_number_literal(slice.lower)
+            upper = (
+                z3.Length(value) if slice.upper is None else self._parse_number_literal(slice.upper)
+            )
+            step = 1 if slice.step is None else self._parse_number_literal(slice.step)
+
+            if not (
+                isinstance(lower, int)
+                and isinstance(upper, (int, z3.ArithRef))
+                and isinstance(step, int)
+            ):
+                raise Z3ParseException(f"Invalid slicing indexes {lower}, {upper}, {step}")
+
+            if step == 1:
+                return z3.SubString(value, lower, upper - lower)
+
+            # unhandled case: the upper bound is indeterminant
+            if step != 1 and upper == z3.Length(value):
+                raise Z3ParseException(
+                    "Unable to convert a slicing operation with a non-unit step length and an indeterminant upper bound"
+                )
+
+            return z3.Concat(*(z3.SubString(value, i, 1) for i in range(lower, upper, step)))
+
+        raise Z3ParseException(f"Unhandled subscript operator type {slice}")
+
+    def parse_arguments(self, node: astroid.Arguments) -> Dict[str, z3.ExprRef]:
+        """Convert an astroid Arguments node's parameters to z3 variables."""
+        z3_vars = {}
+
+        annotations = node.annotations
+        arguments = node.args
+        for ann, arg in zip(annotations, arguments):
+            if ann is None:
+                continue
+
+            inferred = safe_infer(ann)
+            if inferred is None or not isinstance(inferred, astroid.ClassDef):
+                continue
+
+            self.types[arg.name] = inferred.name
+
+            if arg.name in self.types and self.types[arg.name] in {"int", "float", "bool", "str"}:
+                z3_vars[arg.name] = self.reduce(arg)
+
+        return z3_vars