+++ CHANGELOG.md
@@ -70,6 +70,19 @@ Custom checkers:
 - Fixed minor typo in an error message in `python_ta/cfg/visitor.py`
 - Updated `ExprWrapper` to support `set/list/tuple` literals and `in/not in` operators
 - Updated `snapshot.py` and `test_snapshot.py` to align with MemoryViz 0.2.0 updates
+- Updated `ExprWrapper` to support string variables and `==`, `in/not in`, indexing and slicing operators
+- Added protected `_z3_vars` attribute to `ControlFlowGraph` to store variables to be used in Z3 solver
+- Removed unused imports from `python_ta/cfg/graph.py`
+- Extended functionality of `ExprWrapper` class to include function definitions' arguments and name assignments
+- Added `z3` to dependencies installed as part of the `docs` job in the GitHub Actions workflow
+- Added tests to maintain/increase coverage of `visitor.py`, `graph.py`, and `ExprWrapper.py`
+- Removed deprecated and redundant `future` argument from `node.frame()` call in `invalid_name_checker.py`
+- Updated pylint to v3.2.6 and astroid to v3.2.4 (no new checks were enabled by default)
+- Excluded `node_modules/` folder from package autodiscovery
+- Updated `graph.py` to augment control flow graph edges with z3 constraints
+- Added support for the `!=` operator and replaced dictionary indexing with `.get` in `ExprWrapper`.
+- Refactored `Z3Visitor` to use `safe_infer()` instead of `inferred()` and added handling of `AstroidError`.
+- Add `negate` attribute to `CFGEdge`
 
 ## [2.7.0] - 2024-12-14